+++ src/Component/Dto/GameDto.php
@@ -5,6 +5,7 @@ use App\Component\Type\GameState;
 class GameDto
 {
     public string $id;
+    public string $gameCode;
     
     public GameState $playState;
     public float $thinkTime;