+++ src/Component/Manager/Games/ChessGameManager.php
@@ -15,6 +15,7 @@ use App\Component\Manager\BoardGameManager;
 use App\Component\Websocket\Client\WebsocketClientInterface;
 use App\Component\Rules\BoardGame\Game;
 use App\Component\Rules\BoardGame\Player;
+use App\Component\Rules\BoardGame\ChessMove;
 use App\Component\AI\EngineFactory as AiEngineFactory;
 use App\Component\Utils\Guid;
 use App\Component\Websocket\WebSocketState;
@@ -32,6 +33,7 @@ use App\Component\Dto\Actions\ConnectionInfoActionDto;
 use App\Component\Dto\Actions\GameRestoreActionDto;
 use App\Component\Dto\Actions\GameCreatedActionDto;
 use App\Component\Dto\Actions\ChessGameStartedActionDto;
+use App\Component\Dto\Actions\GameEndedActionDto;
 use App\Component\Dto\Actions\HintMovesActionDto;
 use App\Component\Dto\Actions\ChessMoveMadeActionDto;
 use App\Component\Dto\Actions\UndoActionDto;
@@ -70,7 +72,11 @@ final class ChessGameManager extends BoardGameManager
                 if ( $this->Game->CurrentPlayer == PlayerColor::White ) {
                     $promise = Async\async( function () {
                         $this->logger->log( "GameManager CurrentPlayer: White", 'GameManager' );
-                        $this->EnginMoves( $this->Clients->get( PlayerColor::Black->value ) );
+                        
+                        $move = $this->Engine->GetBestMove();
+                        //$this->logger->log( 'Engine Best Move: ' . print_r( $move, true ), 'EnginMoves' );
+                        
+                        $this->EnginMoves( $this->Clients->get( PlayerColor::Black->value ), $move );
                     })();
                     \React\Async\await( $promise );
                 }
@@ -247,6 +253,8 @@ final class ChessGameManager extends BoardGameManager
             $this->Game->ThinkStart = new \DateTime( 'now' );
             $this->DoDoubling();
             $this->Game->SwitchPlayer();
+            
+            $otherSocket = $this->Game->OtherPlayer();
             $this->Send( $otherSocket, $action );
         } else if ( $actionName == ActionNames::requestHint ) {
             if ( ! $this->Game->IsGoldGame && $this->Game->CurrentPlayer == PlayerColor::Black ) {
@@ -260,7 +268,7 @@ final class ChessGameManager extends BoardGameManager
                 $this->Send( $otherSocket, $action );
             }
         } else if ( $actionName == ActionNames::resign ) {
-            $winner = $this->Clients->get( PlayerColor::Black->value ) == $otherSocket ? PlayerColor::Black : PlayerColor::White;
+            $winner = \in_array( $this->Clients->get( PlayerColor::Black->value ), $otherSockets ) ? PlayerColor::Black : PlayerColor::White;
             $this->Resign( $winner );
         } else if ( $actionName == ActionNames::exitGame ) {
             $this->logger->log( 'exitGame action recieved from GameManager.', 'GameManager' );
@@ -274,19 +282,26 @@ final class ChessGameManager extends BoardGameManager
     {
         $winner = $this->GetWinner();
         $this->Game->SwitchPlayer();
-        if ( $winner ) {
-            $this->EndGame( $winner );
-        } else {
-           if ( $this->AisTurn() ) {
-                $this->logger->log( "NewTurn for AI", 'SwitchPlayer' );
-                $this->EnginMoves( $socket );
+        if ( $this->AisTurn() ) {
+            $this->logger->log( "NewTurn for AI", 'SwitchPlayer' );
+            
+            $move = $this->Engine->GetBestMove();
+            //$this->logger->log( 'Engine Best Move: ' . print_r( $move, true ), 'EnginMoves' );
+            
+            if ( $winner && ! $move ) {
+                $this->EndGame( $winner );
             }
+            
+            $this->EnginMoves( $socket, $move );
         }
     }
     
     protected function GetWinner(): ?PlayerColor
     {
         $winner = null;
+        if ( $this->Game->UnderCheck ) {
+            $winner = $this->Game->CauseCheckPlayer;
+        }
         
         return $winner;
     }
@@ -307,32 +322,32 @@ final class ChessGameManager extends BoardGameManager
     
     protected function DoMove( ChessMoveMadeActionDto $action ): void
     {
+        if ( ! $action->move ) {
+            return;
+        }
+        
         //$color  = $move->color;
         $move   = Mapper::ChessMoveToChessMove( $action->move, $this->Game );
         $this->Game->MakeMove( $move );
     }
     
-    protected function EnginMoves( WebsocketClientInterface $client ): void
+    protected function EnginMoves( WebsocketClientInterface $client, ?ChessMove $move ): void
     {
-        $move = $this->Engine->GetBestMove();
-        $this->logger->log( 'Engine Best Move: ' . print_r( $move, true ), 'EnginMoves' );
-        
-        if ( ! $move ) {
-            return;
-        }
-        
         $promise = Async\async( function () use ( $client, $move ) {
             $sleepMileseconds   = \rand( 700, 1200 );
             Async\delay( $sleepMileseconds / 1000 );
             
-            $moveDto = Mapper::ChessMoveToDto( $move );
-            $moveDto->animate = true;
             $dto = new ChessOpponentMoveActionDto();
-            $dto->move = $moveDto;
-            
-            $this->Game->MakeMove( $move );
+            if ( $move ) {
+                $this->Game->MakeMove( $move );
+                $moveDto = Mapper::ChessMoveToDto( $move );
+                $moveDto->animate = true;
+                
+                $dto->move = $moveDto;
+            }
             
             $dto->game = Mapper::BoardGameToDto( $this->Game );
+            $dto->moveTimer = Game::ClientCountDown;
             $dto->myColor = $this->Game->CurrentPlayer;
             
             if ( $this->Game->CurrentPlayer == PlayerColor::Black ) {