+++ src/Component/Rules/BoardGame/ChessRules.php
@@ -17,12 +17,17 @@ class ChessRules
     private  $logger;
     
     /** @var Collection | ChessSquare[] */
-    private $gameSquares;
+    public $gameSquares;
     
     public function __construct( ChessGame $game, GameLogger $logger )
     {
         $this->game             = $game;
         $this->logger           = $logger;
+    }
+    
+    public function DebugLegalMoves( string $cellKey ): Collection
+    {
+        return $this->GetLegalMoves( $this->gameSquares[$cellKey] );
     }
     
     // Return true if the given side is checkmate
@@ -78,10 +83,18 @@ class ChessRules
                 
             case ChessPieceType::Queen:	// Queen piece
                 $this->GetQueenMoves( $source, $LegalMoves );
+                
+//                 $debugLegalMoves = \print_r( $LegalMoves->toArray(), true );
+//                 $this->logger->log( "Queen Valid Moves: {$debugLegalMoves}", 'EnginMoves' );
+                
                 break;
                 
             case ChessPieceType::King:	// king piece
                 $this->GetKingMoves( $source, $LegalMoves );
+                
+//                 $debugLegalMoves = \print_r( $LegalMoves->toArray(), true );
+//                 $this->logger->log( "King Valid Moves: {$debugLegalMoves}", 'EnginMoves' );
+                
                 break;
         }
         
@@ -92,6 +105,10 @@ class ChessRules
     public function GetLegalMoves( ChessSquare $source ): Collection
     {
         $LegalMoves = $this->GetPossibleMoves( $source );	// Get the legal moves
+        
+        //$debugLegalMoves = \print_r( $LegalMoves->toArray(), true );
+        //$this->logger->log( "Get the legal moves: {$debugLegalMoves}", 'EnginMoves' );
+        
         $ToRemove = new ArrayCollection();	// contains a list of all the moves to remove
         
         // Now check and mark all the moves which moves user under check
@@ -100,6 +117,10 @@ class ChessRules
             $move = new ChessMove();
             $move->From = $source;
             $move->To = $target;
+            $move->Piece = $source->Piece;
+            $move->Type = $target->Piece ? ChessMoveType::CaputreMove : ChessMoveType::NormalMove;
+            $move->CapturedPiece = $target->Piece ?: null;
+            
             if ( $this->CauseCheck( $move ) ) {
                 $ToRemove[] = $target;
             }
@@ -117,6 +138,9 @@ class ChessRules
             }
         }
         
+        //$debugLegalMoves = \print_r( $ToRemove->toArray(), true );
+        //$this->logger->log( "To Remove Legal Moves: {$debugLegalMoves}", 'EnginMoves' );
+        
         // remove all the illegal moves
         foreach ( $ToRemove as $cell ) {
             $LegalMoves->removeElement( $cell );	// remove the illegal move
@@ -132,7 +156,7 @@ class ChessRules
         $this->actualGameSquares();
         
         $TotalMoves = new ArrayCollection();
-        $PlayerCells = $this->game->GetSideCell( $PlayerSide->type );
+        $PlayerCells = $this->game->GetSideCell( $PlayerSide->type, $this->gameSquares );
         
         // Loop all the owner squars and get their possible moves
         foreach ( $PlayerCells as $CellName ) {
@@ -169,8 +193,10 @@ class ChessRules
     // Generate all the good capture moves. These are used for quiescent searching
     public function GenerateGoodCaptureMoves( ChessSide $PlayerSide ): Collection
     {
+        $this->actualGameSquares();
+        
         $TotalMoves = new ArrayCollection();
-        $PlayerCells = $this->game->GetSideCell( $PlayerSide );
+        $PlayerCells = $this->game->GetSideCell( $PlayerSide, $this->gameSquares );
         
         // Loop all the owner squars and get their possible moves
         foreach ( $PlayerCells as $CellName ) {
@@ -369,12 +395,14 @@ class ChessRules
     }
     
     // Return true if the given side type is under check state
-    public function IsUnderCheck( PlayerColor $PlayerSide ): bool
+    public function IsUnderCheck( PlayerColor $PlayerSide, bool $actualizeGameSquares = true ): bool
     {
-        $this->actualGameSquares();
+        if ( $actualizeGameSquares ) {
+            $this->actualGameSquares();
+        }
         
         $OwnerKingCell=null;
-        $OwnerCells = $this->game->GetSideCell( $PlayerSide );
+        $OwnerCells = $this->game->GetSideCell( $PlayerSide, $this->gameSquares );
         
         // loop all the owner squars and get his king cell
         foreach ( $OwnerCells as $CellName ) {
@@ -384,18 +412,22 @@ class ChessRules
             }
         }
         
+        /* This Create Infinite recursion? at ArrayCollection.php
+         * =======================================================
         // Loop all the enemy squars and get their possible moves
-        $EnemyCells = $this->game->GetSideCell( ( new ChessSide( $PlayerSide ) )->Enemy() );
+        $EnemyCells = $this->game->GetSideCell( ( new ChessSide( $PlayerSide ) )->Enemy(), $this->gameSquares );
         //$this->logger->log( "Game Squares: " . \print_r( $this->gameSquares, true ), 'EnginMoves' );
         foreach ( $EnemyCells as $CellName ) {
             //$this->logger->log( "Player Side: {$PlayerSide->value}", 'EnginMoves' );
             //$this->logger->log( "Cell Name: {$CellName}", 'EnginMoves' );
             $moves = $this->GetPossibleMoves( $this->gameSquares[$CellName] );	// Get the moves for the enemy piece
+            
             // King is directly under attack
             if ( $moves->contains( $OwnerKingCell ) ) {
                 return true;
             }
         }
+        */
         
         return false;
     }
@@ -404,7 +436,7 @@ class ChessRules
     public function AnalyzeBoard( PlayerColor $PlayerSide ): int
     {
         $Score = 0;
-        $OwnerCells = $this->game->GetSideCell( $PlayerSide );
+        $OwnerCells = $this->game->GetSideCell( $PlayerSide, $this->gameSquares );
         
         // loop all the owner squars and get his king cell
         foreach ( $OwnerCells as $ChessCell ) {
@@ -435,12 +467,11 @@ class ChessRules
     
     private function actualGameSquares(): void
     {
-        $this->gameSquares = new ArrayCollection();
-        $gameSquaresClone = $this->game->Squares->getValues();
-        
-        foreach ( $gameSquaresClone as $suqare ) {
-            $this->gameSquares->set( "{$suqare}", clone $suqare );
-        }
+        $this->gameSquares = $this->game->Squares->map(
+            function( $entry ) {
+                return clone $entry;
+            }
+        );
     }
     
     // return type of the move given the move object
@@ -576,7 +607,7 @@ class ChessRules
         $TotalMoves = 0;
         
         // Loop all the owner squars and get their possible moves
-        $PlayerCells = $this->game->GetSideCell( $PlayerSide );
+        $PlayerCells = $this->game->GetSideCell( $PlayerSide, $this->gameSquares );
         foreach ( $PlayerCells as $CellName ) {
             $moves = $this->GetLegalMoves( $this->gameSquares[$CellName] );	// Get all the legal moves for the owner piece
             $TotalMoves += $moves->count();
@@ -596,7 +627,9 @@ class ChessRules
         
         // To check if a move cause check, we actually need to execute and check the result of that move
         $this->ExecuteMove( $move );
-        $CauseCheck = $this->IsUnderCheck( $PlayerSide );
+//         $debugMove = \print_r( $move, true );
+//         $this->logger->log( "Cause Check Move: {$debugMove}", 'EnginMoves' );
+        $CauseCheck = $this->IsUnderCheck( $PlayerSide, false );
         $this->UndoMove( $move );	// undo the move
         
         return $CauseCheck;
@@ -605,7 +638,8 @@ class ChessRules
     // calculate the possible moves for the pawn object and insert them into passed array
     private function GetPawnMoves( ChessSquare $source, Collection &$moves ): void
     {
-        if ( $source->Piece->Side->isWhite() ) {
+        //if ( $source->Piece->Side->isWhite() ) {
+        if ( $source->Piece->Side->isBlack() ) { // In my Board The ranks are increasing from White to Black
             // Calculate moves for the white piece
             $newcell = $this->TopCell( $source );
             if ( $newcell && ! $newcell->Piece ) { // Top cell is available for the move
@@ -894,6 +928,7 @@ class ChessRules
         
         // Check all the move squars available in bottom-left direction
         $newcell = $this->BottomLeftCell( $source );
+        //$this->logger->log( "Bishop BottomLeftCell: " . \print_r( $newcell, true ), 'EnginMoves' );
         while ( $newcell != null ) {	// move as long as cell is available in this direction
             if ( ! $newcell->Piece ) {	//next cell is available for move
                 $moves[] = $newcell;