+++ themes/GamePlatform_VelzonSaas/assets/js/games/application/dto/Actions/chessOpponentMoveActionDto.ts
@@ -5,10 +5,11 @@ import BoardGameDto from '_@/GamePlatform/Model/BoardGame/gameDto';
 import PlayerColor from '_@/GamePlatform/Model/BoardGame/playerColor';
 
 interface ChessOpponentMoveActionDto extends ActionDto {
-    move: ChessMoveDto;
+    move?: ChessMoveDto;
     myColor: PlayerColor;
     
     game?: BoardGameDto;
+    moveTimer?: number;
 }
 
 export default ChessOpponentMoveActionDto;