+++ themes/GamePlatform_VelzonSaas/assets/js/games/application/services/websocket/chess.service.ts
@@ -1,5 +1,6 @@
 import { Injectable, Inject, Injector } from '@angular/core';
 import { AbstractGameService } from './abstract-game.service';
+import {NgxChessBoardService} from 'ngx-chess-board';
 
 // NGRX Store
 import { loadGameRooms } from '../../+store/game.actions';
@@ -36,6 +37,7 @@ export class ChessService extends AbstractGameService
 {
     constructor(
         @Inject( Injector ) private injector: Injector,
+        //@Inject( NgxChessBoardService ) private ngxChessBoardService: NgxChessBoardService,
     ) {
         super( injector );
     }
@@ -89,8 +91,7 @@ export class ChessService extends AbstractGameService
         
         //console.log( 'User in State', this.appState.user );
         if ( this.appState.user.getValue() ) {
-            //this.statusMessageService.setWaitingForConnect();
-            this.statusMessageService.setNotGameStarted();
+            this.statusMessageService.setWaitingForConnect();
         } else {
             this.statusMessageService.setNotLoggedIn();
             this.appState.hideBusy();
@@ -144,11 +145,14 @@ export class ChessService extends AbstractGameService
                 const action = JSON.parse( message.data ) as ChessGameStartedActionDto;
                 console.log( 'WebSocket Action Chess Game Started', action );
                 
-                this.appState.boardGame.setValue({
+                const cGame = {
                     ...game,
-                    currentPlayer: action.game.currentPlayer
-                });
+                    currentPlayer: action.game.currentPlayer,
+                    playState: action.game.playState
+                };
                 
+                this.appState.boardGame.setValue( cGame );
+                this.statusMessageService.setTextMessage( cGame );
                 this.appState.moveTimer.setValue( action.moveTimer );
                 
                 break;
@@ -225,14 +229,22 @@ export class ChessService extends AbstractGameService
                 const action = JSON.parse( message.data ) as ChessOpponentMoveActionDto;
                 console.log( 'WebSocket Action Opponent Move ', action );
                 
-                this.doMove( action.move );
-                
-                break;
-            }
-            case ActionNames.undoMove: {
-                //console.log( 'WebSocket Action Undo Move', action.actionName );
-                
-                this.undoMove();
+                if ( action.move ) {
+                    this.doMove( action.move );
+                } else {
+                    //this.ngxChessBoardService.moveChange.emit();
+                    
+                    const cGame = {
+                        ...game,
+                        currentPlayer: game.currentPlayer === PlayerColor.black ? PlayerColor.white : PlayerColor.black
+                    };
+                    
+                    this.appState.boardGame.setValue( cGame );
+                    this.statusMessageService.setTextMessage( cGame );
+                    if ( action.moveTimer ) {
+                        this.appState.moveTimer.setValue( action.moveTimer );
+                    }
+                }
                 
                 break;
             }
@@ -298,45 +310,24 @@ export class ChessService extends AbstractGameService
         this.chesUserMoves = [];
     }
     
-    /*
-    doOpponentMove( move: ChessMoveDto ): void
-    {
-        const game = this.appState.boardGame.getValue();
-        const gameClone = JSON.parse( JSON.stringify( game ) ) as BoardGameDto;
-        const isWhite = move.color === PlayerColor.white;
-        const from = isWhite ? 25 - move.from : move.from;
-        const to = isWhite ? 25 - move.to : move.to;
-        //const checker = <CheckerDto>gameClone.points[from].checkers.pop();
-        
-        // hitting opponent checker
-        const hit = gameClone.points[to].checkers.find(
-            ( c ) => c.color !== move.color
-        );
-        if ( hit ) {
-            gameClone.points[to].checkers.pop();
-            const barIdx = isWhite ? 0 : 25;
-            gameClone.points[barIdx].checkers.push( hit );
-        }
-        
-        //gameClone.points[to].checkers.push( checker );
-        
-        this.appState.boardGame.setValue( gameClone );
-    }
-    */
-    
     doMove( move: ChessMoveDto ): void
     {
         //console.log( 'Chess Moee', move );
         this.chesUserMoves.push( { ...move, nextMoves: [] } ); // server does not need to know nextMoves.
+        const myColor = this.appState.myColor.getValue();
+        const oponent = myColor === PlayerColor.black ? PlayerColor.white : PlayerColor.black
+        
         const prevGame = this.appState.boardGame.getValue();
         this.gameHistory.push( prevGame );
         
+        /*
         const gameClone = JSON.parse( JSON.stringify( prevGame ) ) as BoardGameDto;
         gameClone.currentPlayer = move.color;
         this.appState.boardGame.setValue( gameClone );
+        */
         
-        const myColor = this.appState.myColor.getValue();
-        const oponent = myColor === PlayerColor.black ? PlayerColor.white : PlayerColor.black
+        //alert( `DoMove Oponent: ${oponent}` );
+        //alert( `DoMove Player: ${move.color}` );
         if ( move.color === oponent ) {
             this.appState.chessOpponentMove.setValue( move );
         }
@@ -442,5 +433,18 @@ export class ChessService extends AbstractGameService
         this.appState.moveTimer.setValue( 40 );
         this.sendMessage( JSON.stringify( action ) );
         this.statusMessageService.setWaitingForDoubleResponse();
+    }
+    
+    switchPlayer(): void
+    {
+        const game = this.appState.boardGame.getValue();
+        const cGame = {
+            ...game,
+            currentPlayer: game.currentPlayer === PlayerColor.black ? PlayerColor.white : PlayerColor.black
+        };
+        
+        this.appState.boardGame.setValue( cGame );
+        this.statusMessageService.setTextMessage( cGame );
+        this.appState.moveTimer.setValue( 40 );
     }
 }